fix suppression des ennemis

+++ src/sfml/PersoSFML.cpp
@@ -1,4 +1,6 @@
 #include "PersoSFML.h"
+#include "Config.h"
+
 #include <iostream>
 
 PersoSFML::PersoSFML()
@@ -14,9 +16,10 @@ void PersoSFML::charge_perso(Perso * p,const sf::Texture& texture,const int &tai
     persosfml.setScale(p->get_taille()*taille_case/texture.getSize().x, p->get_taille()*taille_case/texture.getSize().y);
 }
 
-void PersoSFML::mettre_a_jour(const int &taille_case, const int &x0, const int &y0)
+void PersoSFML::mettre_a_jour(const float &taille_case, const int &x0, const int &y0)
 {
     persosfml.setPosition(perso->get_pos_x() * taille_case + x0, perso->get_pos_y() * taille_case + y0);
+    //std::cout << perso->get_pos_x() * taille_case + x0 << " " << perso->get_pos_y() * taille_case + y0 << std::endl;
     persosfml.setRotation(-perso->get_orientation_degre()+ 180);
 }
